+++ tests/test_exporter.py
@@ -1,10 +1,13 @@
 import json
 import os
+from pathlib import Path
 
 import pytest
 from osgeo import ogr
 
+from threedigrid.admin.exporters.geopackage import GeopackageExporter
 from threedigrid.admin.exporters.geopackage.exporter import GpkgExporter
+from threedigrid.admin.fragments.exporters import FragmentsOgrExporter
 from threedigrid.admin.lines.exporters import LinesOgrExporter
 
 test_file_dir = os.path.join(os.getcwd(), "tests/test_files")
@@ -34,6 +37,68 @@ def test_nodes_gpgk_export(ga, tmp_path):
     s = ogr.Open(path)
     layer = s.GetLayer("nodes")
     assert layer.GetFeatureCount() == nodes_2d_open_water.id.size
+
+
+def test_fragments_gpgk_export(ga_fragments, tmp_path):
+    path = str(tmp_path / ("exporter_test_fragments.gpkg"))
+    exporter = FragmentsOgrExporter(ga_fragments.fragments)
+    exporter.save(path, "fragments", "4326")
+    assert os.path.exists(path)
+    s = ogr.Open(path)
+    layer = s.GetLayer()
+    assert layer.GetFeatureCount() == (
+        ga_fragments.fragments.id.size - 1
+    )  # minus dummy
+
+
+def test_fragments_to_gpgk_export(ga_fragments, tmp_path):
+    path = str(tmp_path / "only_fragment_export.gpkg")
+    ga_fragments.fragments.to_gpkg(path, "fragments")
+    s = ogr.Open(path)
+    layer = s.GetLayer("fragments")
+    assert layer.GetFeatureCount() == (
+        ga_fragments.fragments.id.size - 1
+    )  # minus dummy
+
+
+def test_fragments_to_geojson_export(ga_fragments, tmp_path):
+    path = str(tmp_path / "only_fragment_export.json")
+    ga_fragments.fragments.to_geojson(path, use_ogr=True)
+    s = ogr.Open(path)
+    layer = s.GetLayer(Path(path).name)
+    assert layer.GetFeatureCount() == (
+        ga_fragments.fragments.id.size - 1
+    )  # minus dummy
+    with open(path) as file:
+        data = json.load(file)
+        assert len(data["features"][0]["properties"]) == 2
+        assert "node_id" in data["features"][0]["properties"]
+
+    path = str(tmp_path / "only_fragment_export_2.json")
+    ga_fragments.fragments.to_geojson(path, use_ogr=False)
+    s = ogr.Open(path)
+    layer = s.GetLayer(Path(path).stem)
+    # We export more properties and use different layer name
+    assert layer.GetFeatureCount() == (
+        ga_fragments.fragments.id.size - 1
+    )  # minus dummy
+    with open(path) as file:
+        data = json.load(file)
+        assert len(data["features"][0]["properties"]) == 3
+        assert "node_id" in data["features"][0]["properties"]
+        assert "model_type" in data["features"][0]["properties"]
+
+
+def test_fragments_gpkg_conversion(ga_fragments, ga_fragment_path, tmp_path):
+    exporter = GeopackageExporter(
+        ga_fragment_path, str(tmp_path / "fragment_export.gpkg")
+    )
+    exporter.export()
+    s = ogr.Open(str(tmp_path / "fragment_export.gpkg"))
+    layer = s.GetLayer("fragment")
+    assert layer.GetFeatureCount() == (
+        ga_fragments.fragments.id.size - 1
+    )  # minus dummy
 
 
 def test_meta_data_gpgk_export(ga, tmp_path):
@@ -138,6 +203,7 @@ def test_export_null(ga, tmp_path):
         ("grid", "grid_2D_open_water"),
         ("grid", "grid_2D_groundwater"),
         ("flowlines", "flowlines"),
+        ("fragments", "fragments"),
     ],
 )
 def test_export_method(ga_export, export_method, expected_filename):